Add a basic log (etc.) archive uploader for Drone builds
This is a simple plugin that will let us upload artefacts from Drone
builds to an S3 bucket.

It really differs only from the drone-s3-bash-cache plugin we're already
using in that it lacks a "restore" function, and has a slightly
different upload naming scheme. These differences could possibly be
abstracted away to configuration, but for the time being I just want to
get something simple working.